+++ backend/src/services/methods.rs
@@ -160,3 +160,111 @@ pub async fn fetch_nodes<
 
     Ok(nodes)
 }
+
+/// Return `VDisk` information by id
+///
+/// # Errors
+///
+/// This function will return an error if vdisks information couldn't be fetched or no vdisk with
+/// provided id was found
+pub async fn get_vdisk_by_id(client: &HttpBobClient, vdisk_id: u64) -> AxumResult<VDisk> {
+    let virtual_disks = fetch_vdisks(client.api_main()).await?;
+    let virtual_disks = virtual_disks
+        .iter()
+        .find(|vdisk| vdisk.id as u64 == vdisk_id)
+        .ok_or_else(|| StatusCode::NOT_FOUND.into_response())?;
+    let clients = virtual_disks
+        .replicas
+        .iter()
+        .flatten()
+        .map(|replica| replica.node.clone())
+        .collect::<HashSet<_>>()
+        .iter()
+        .filter_map(|node_name| client.api_secondary(node_name))
+        .collect::<Vec<_>>();
+    let partition_count = if let Some(handle) = clients.first() {
+        handle.get_partitions(vdisk_id as i32).await.map_or_else(
+            |_err| 0,
+            |parts| {
+                if let GetPartitionsResponse::NodeInfoAndJSONArrayWithPartitionsInfo(parts) = parts
+                {
+                    parts.partitions.unwrap_or_default().len()
+                } else {
+                    0
+                }
+            },
+        )
+    } else {
+        0
+    };
+    let mut disks: FuturesUnordered<_> = clients
+        .iter()
+        .map(move |&node| {
+            let handle = node.clone();
+            tokio::spawn(async move { (handle.get_status().await, handle.get_disks().await) })
+        })
+        .collect();
+    let mut replicas: HashMap<_, _> = virtual_disks
+        .replicas
+        .clone()
+        .into_iter()
+        .flatten()
+        .map(|replica| {
+            (
+                (replica.disk.clone(), replica.node.clone()),
+                Replica {
+                    node: replica.node,
+                    disk: replica.disk,
+                    path: replica.path,
+                    status: ReplicaStatus::Offline(vec![ReplicaProblem::NodeUnavailable]),
+                },
+            )
+        })
+        .collect();
+    while let Some(res) = disks.next().await {
+        if let Ok((
+            Ok(GetStatusResponse::AJSONWithNodeInfo(status)),
+            Ok(GetDisksResponse::AJSONArrayWithDisksAndTheirStates(disks)),
+        )) = res
+        {
+            for disk in disks {
+                replicas.insert(
+                    (disk.name.clone(), status.name.clone()),
+                    Replica {
+                        node: status.name.clone(),
+                        disk: disk.name,
+                        path: disk.path,
+                        status: disk
+                            .is_active
+                            .then_some(ReplicaStatus::Good)
+                            .unwrap_or_else(|| {
+                                ReplicaStatus::Offline(vec![ReplicaProblem::DiskUnavailable])
+                            }),
+                    },
+                );
+            }
+        } else {
+            tracing::warn!("couldn't receive node's space info");
+        }
+    }
+
+    let replicas: Vec<_> = replicas.into_values().collect();
+    let count = replicas
+        .iter()
+        .filter(|replica| matches!(replica.status, ReplicaStatus::Offline(_)))
+        .count();
+    let status = if count == 0 {
+        VDiskStatus::Good
+    } else if count == replicas.len() {
+        VDiskStatus::Offline
+    } else {
+        VDiskStatus::Bad
+    };
+
+    Ok(VDisk {
+        id: vdisk_id,
+        status,
+        partition_count: partition_count as u64,
+        replicas,
+    })
+}
